requirements.txt
chromadb
pypdf
langchain_community
langchain_text_splitters
langchain_google_genai
langchain_core
gTTS
PyAudio
SpeechRecognition
pydub
openai
pyautogui
google
duckduckgo_search
openai-whisper
setuptools-rust
RealtimeSTT
pyttsx3